refactor() : product 클래스 패키지 변경 - persistence 아래 entity 패키지로 이동
